+++ packages/apps/marmalade-marketplace/src/graphql/queries/getTokens.ts
@@ -1,0 +1,15 @@
+export const GET_TOKENS_QUERY = `
+  query GetTokens($accountName: String!) {
+    nonFungibleAccount(accountName: $accountName) {
+      accountName
+      id
+      nonFungibleTokenBalances {
+        balance
+        tokenId
+        info {
+          uri
+        }
+      }
+    }
+  }
+`;