+++ packages/apps/marmalade-marketplace/src/pages/index.tsx
@@ -1,10 +1,8 @@
-import Layout from '@/components/Layout';
 import { getSales } from "@/hooks/getSales";
 import { Button, Grid, GridItem, Link } from "@kadena/react-ui";
 import { Token } from "@/components/Token";
 import { useEffect, useState } from "react";
 import { actionBarClass, actionBarSaleActiveClass, actionBarSaleClass } from "@/styles/home.css";
-import { Providers } from "@/providers/Providers";
 
 export default function Home() {
 
@@ -26,46 +24,42 @@ export default function Home() {
   }, [saleStatus]);
 
   return (
-    <Providers>
-      <Layout>
-        <div style={{ marginTop: "100px" }}>
-          <div className={actionBarClass}>
-            <div className={actionBarSaleClass}>
-              <Button
-                className={saleStatus === "CREATED" ? actionBarSaleActiveClass : ""}
-                onClick={() => setSaleStatus("CREATED")}>Active sales</Button>
-              <Button
-                className={saleStatus === "SOLD" ? actionBarSaleActiveClass : ""}
-                onClick={() => setSaleStatus("SOLD")}>Past sales</Button>
-            </div>
-            <Link variant="primary">
-              Sell Token
-            </Link>
-          </div>
-
-          {error && <div>Error: <pre>{JSON.stringify(error, null, 2)}</pre></div>}
-          {loading && <h2>Loading..</h2>}
-
-          <div>
-            <Grid
-              columns={{
-                lg: 4,
-                md: 3,
-                sm: 2,
-                xs: 1,
-              }}
-              gap="xl">
-              {data.map((sale, index) => (
-                <GridItem key={index}>
-                  <a href={`/tokens/${sale.tokenId}`}>
-                    <Token tokenId={sale.tokenId} chainId={sale.chainId} sale={sale} />
-                  </a>
-                </GridItem>
-              ))}
-            </Grid>
-          </div>
+    <div>
+      <div className={actionBarClass}>
+        <div className={actionBarSaleClass}>
+          <Button
+            className={saleStatus === "CREATED" ? actionBarSaleActiveClass : ""}
+            onClick={() => setSaleStatus("CREATED")}>Active sales</Button>
+          <Button
+            className={saleStatus === "SOLD" ? actionBarSaleActiveClass : ""}
+            onClick={() => setSaleStatus("SOLD")}>Past sales</Button>
         </div>
-      </Layout>
-    </Providers>
+        <Link variant="primary">
+          Sell Token
+        </Link>
+      </div>
+
+      {error && <div>Error: <pre>{JSON.stringify(error, null, 2)}</pre></div>}
+      {loading && <h2>Loading..</h2>}
+
+      <div>
+        <Grid
+          columns={{
+            lg: 4,
+            md: 3,
+            sm: 2,
+            xs: 1,
+          }}
+          gap="xl">
+          {data.map((sale, index) => (
+            <GridItem key={index}>
+              <a href={`/tokens/${sale.tokenId}`}>
+                <Token tokenId={sale.tokenId} chainId={sale.chainId} sale={sale} />
+              </a>
+            </GridItem>
+          ))}
+        </Grid>
+      </div>
+    </div>
   );
 }