+++ packages/apps/tools/src/components/Global/Ace/helper.ts
@@ -1,7 +1,7 @@
 export const keyboards = ['vim', 'emacs', 'ace'] as const;
 export type KeyboardHandler = (typeof keyboards)[number];
 
-export const themes = ['monokai', 'one_dark'] as const;
+export const themes = ['monokai', 'one_dark', 'github', 'github_dark'] as const;
 export type Theme = (typeof themes)[number];
 
 export const modes = ['lisp', 'clojure'] as const;