+++ packages/libs/toolbox/src/prelude/preludes/index.ts
@@ -1,2 +1,0 @@
-export { default as chainWebPrelude } from './kadena/chainweb';
-export { default as marmaladePrelude } from './kadena/marmalade';