+++ packages/apps/dev-wallet/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @kadena/dev-wallet
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies \[6a969ebdc]
+- Updated dependencies \[b541021f4]
+- Updated dependencies \[e648c32fa]
+- Updated dependencies \[9c145f196]
+- Updated dependencies \[9c145f196]
+  - @kadena/react-ui\@0.11.0
+  - @kadena/client\@1.12.0
+  - @kadena/client-utils\@0.7.0
+  - @kadena/pactjs-generator\@1.12.0
+
 ## 0.1.1
 
 ### Patch Changes