+++ packages/libs/react-ui/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @kadena/react-ui
 
+## 0.11.0
+
+### Minor Changes
+
+- 6a969ebdc: Add radiogroup component
+- e648c32fa: - Reintroduced `items` prop for Tabs component to offer Dynamic
+  collection support.
+  - Added `tabPanelClassName` prop to Tabs component to forward styles to
+    TabPanel component.
+  - Added export of ITabNode type.
+
+### Patch Changes
+
+- b541021f4: Readme update
+
 ## 0.10.0
 
 ### Minor Changes